requirements.txt
torch
torchvision
pandas
scikit-learn
Pillow
fastapi
uvicorn